test(router): isolate navigation diagnostics E2E
The shared App Router Playwright server enabled navigation diagnostics for every browser test, removing default-disabled coverage and relying on POSIX-only inline environment syntax.

Move the lifecycle assertion to a dedicated Playwright project and server, pass the flag through Playwright environment configuration, and run the project independently in CI.
